+++ lib/components/TheMain.vue
@@ -1,9 +1,9 @@
 <template>
   <TransitionFadeSlide>
     <div
+      :key="showAside"
       class="container clearfix"
       :class="containerClass"
-      :key="showAside"
     >
       <main class="main">
         <TransitionFadeSlide>
@@ -55,7 +55,7 @@ export default {
     },
 
     showAside () {
-      if (this.$page.frontmatter.hasOwnProperty('aside')) {
+      if ('aside' in this.$page.frontmatter) {
         return this.$page.frontmatter.aside
       }
       return true