+++ lib/components/TransitionFadeSlide.vue
@@ -31,6 +31,16 @@ export default {
     },
   },
 
+  computed: {
+    name () {
+      return `fade-slide-${this.direction}`
+    },
+
+    component () {
+      return this.group ? 'transition-group' : 'transition'
+    },
+  },
+
   render (h, { props, children }) {
     return h(
       props.group ? 'TransitionGroup' : 'Transition',
@@ -42,16 +52,6 @@ export default {
       },
       children,
     )
-  },
-
-  computed: {
-    name () {
-      return `fade-slide-${this.direction}`
-    },
-
-    component () {
-      return this.group ? 'transition-group' : 'transition'
-    },
   },
 }
 </script>